Added dev container configuration and updated the docker-compose file to replace missing values with the correct env values. Updated .env.sample to reflect this. Copy .env.sample to your local .env file (create if a .env file does not exist).
Working: The dev container launches and runs the app and db locally. All
good with this.
Does not work: VS Code does not seem to have options to debug the code
for some reason. Ideally, we need to be able to attach a debugger to the
running instance directly from the IDE.